fix codacy warnings (#158)
* fix codacy warnings

* fix tests

* autopep

* fix more warnings

* fix warnings

* fix warnings

* fix warning

+++ utils/oaanalyse.cpp
@@ -197,7 +197,7 @@ int main (int argc, char *argv[]) {
         if (doj) {
                 for (unsigned int ii = 0; ii < arraylist->size (); ii++) {
                         if (ii % 5000 == 0 || verbose >= 3)
-                                printf ("  calculation (j-filter) for array %d/%d\n", ii, (int)arraylist->size ());
+                                printf ("  calculation (j-filter) for array %d/%d\n", (int)ii, (int)arraylist->size ());
 
                         bool ret = testarray (results[ii], maxj);
                         if (ret) {
